Add A/B slots for applications
Add partitions and slots that may contain user-specific applications.
The correct partition to be mounted is determined by the corresponding
parent root filesystem in use. A udev rule is being added that
automatically creates a symlink of the application filesystem device.
The partition is also mounted automatically to /mnt/app.

Signed-off-by: Martin Schwan <m.schwan@phytec.de>